Improved build-package.sh to generate source builds

+++ build-package.sh
@@ -6,6 +6,9 @@
 #             After that, ***cd /tmp*** and upload with: dput -f meiga meiga*.changes
 # -rn distro  The same than -r, but precompiling C sources to avoid dependancy on
 #             Vala package in the target distro
+# -s          Prepares a source release
+# -sn         The same than -s, but precompiling C sources to avoid dependancy on
+#             Vala package
 
 # -------------------------------------------------
 
@@ -18,11 +21,13 @@ case $LSB_VENDOR in
   Ubuntu|Debian)
     VALA_INSTALLED=0
     WKDIR=`pwd`
+    MEIGA_VERSION=`grep AC_INIT "$WKDIR/configure.ac" | sed -e 's/[^,]*,\[\([^]]*\).*/\1/'`
+
     if [ "$#" -ge 1 ]
     then
       if [ "$1" = "-d" ]
       then
-	shift
+        shift
         if [ -f /usr/bin/lsb_release ]
         then
           UBUNTU_RELEASE=`lsb_release -r | { read _ X; echo $X; }`
@@ -51,29 +56,46 @@ case $LSB_VENDOR in
         fi
       fi
 
-      if [ "$1" = "-r" -o "$1" = "-rn" ]
+      if [ "$1" = "-r" -o "$1" = "-rn" -o "$1" = "-s" -o "$1" = "-sn" ]
       then
         RELEASETYPE="$1"
-	DISTRO="$2"
-	BUILDPATH="/tmp/meiga"
-	shift
-	shift
-	rm -rf "$BUILDPATH" "$BUILDPATH"_*
-	mkdir "$BUILDPATH" \
+        DISTRO="$2"
+        PACKAGE="meiga-$MEIGA_VERSION"
+        BUILDPATH="/tmp/$PACKAGE"
+
+        shift
+        if [ "$RELEASETYPE" = "-r" -o "$RELEASETYPE" = "-rn" ]
+        then
+          shift
+        fi
+
+        if [ ! -d .git ]
+        then
+          echo "This kind of release can only be done from a git clone, not from downloaded sources"
+          exit
+        fi
+
+        rm -rf "$BUILDPATH" "$BUILDPATH"_*
+        mkdir "$BUILDPATH" \
         && cp -a .git "$BUILDPATH" \
         && cd "$BUILDPATH" \
         && git reset --hard HEAD \
         && rm -rf .git \
-        && sed -e "s/) unstable/$DISTRO) $DISTRO/" < "$WKDIR/debian/changelog" > "$BUILDPATH/debian/changelog"
+        && {
+          if [ "$RELEASETYPE" = "-r" -o "$RELEASETYPE" = "-rn" ]
+          then
+	    sed -e "s/) unstable/$DISTRO) $DISTRO/" < "$WKDIR/debian/changelog" > "$BUILDPATH/debian/changelog"
+          fi
+        }
 
-	BINARYONLY="-S"
+        BINARYONLY="-S"
         SIGNCHANGES=""
 
-        if [ "$RELEASETYPE" = "-rn" ]
+        if [ "$RELEASETYPE" = "-rn" -o "$RELEASETYPE" = "-sn" ]
         then
           # Remove valac dependency from debian/control in /tmp/meiga
           sed -e 's/, valac (.*)//' < "$WKDIR/debian/control" > "$BUILDPATH/debian/control" \
-          && ./autogen.sh \
+          && cd "$BUILDPATH" && ./autogen.sh \
           && cd "$BUILDPATH/src/gui" && make meiga.vala.stamp && cd "$BUILDPATH" \
           && cd "$BUILDPATH/src/server" && make meiga.vala.stamp && cd "$BUILDPATH"
         fi
@@ -99,10 +121,17 @@ case $LSB_VENDOR in
       echo
     fi
 
-    # If you're using a distribution for which an updated Vala package 
-    # exists and you want to check the dependency, remove the "-d" option
+    if [ "$RELEASETYPE" = "-s" -o "$RELEASETYPE" = "-sn" ]
+    then
+      cd "$BUILDPATH" \
+      && cd .. \
+      && tar zcvf "$PACKAGE.tar.gz" "$PACKAGE"
+    else
+      # If you're using a distribution for which an updated Vala package 
+      # exists and you want to check the dependency, remove the "-d" option
 
-    dpkg-buildpackage -d -rfakeroot $BINARYONLY $SIGNCHANGES
+      dpkg-buildpackage -d -rfakeroot $BINARYONLY $SIGNCHANGES
+    fi
 
     if [ "$VALA_INSTALLED" = "1" ]
     then
